Revert " Update substrate version to "polkadot-v0.9.12" (#180)" (#184)
This reverts commit 5c9c2999e68d33b022234d3aed7919e920db4d0d.

+++ pallet/src/lib.rs
@@ -61,6 +61,7 @@ pub mod pallet {
     }
 
     #[pallet::event]
+    #[pallet::metadata(T::AccountId = "AccountId")]
     #[pallet::generate_deposit(pub(super) fn deposit_event)]
     pub enum Event<T: Config> {
         ChangeValidators(Vec<T::AccountId>, u32),
@@ -237,6 +238,6 @@ pub mod pallet {
             Self::update_authorities(authorities.as_slice());
         }
 
-        fn on_disabled(_validator_index: u32) {}
+        fn on_disabled(_validator_index: usize) {}
     }
 }
